pyproject.toml
[project]
name = "mcodex"
version = "0.1.0"
description = "CLI tool for managing text projects."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "docopt>=0.6.2",
  "pyyaml>=6.0.0",
]

[project.scripts]
mcodex = "mcodex.__main__:main"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcodex"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_equality = true
check_untyped_defs = true
pretty = true
show_error_codes = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-q"

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "pytest>=8.0.0",
  "behave>=1.2.6",
  "types-PyYAML>=6.0.0.0",
]

